lessons/lesson-02-multi-file-editing/src/pyproject.toml
[project]
name = "transcript-analytics"
version = "0.1.0"
dependencies = [
    "python-dotenv",
    "google-generativeai",
    "pydantic"
]

[tool.uv]
dev-dependencies = []